Ensure that we highlight dollar-slashy-strings correctly
This had no tests before, and it had been broken.

Fix the logic by always calling the dollar-slashy-string functions
first, but delegate to slashy-string-function where appropriate.

Add a test, and also split out a test from
groovy-highlight-slashy-string for the case when we mix slashy-strings
with $. This is easy to break, so having a separate test makes it
easier to spot what's wrong.

Aggressively comment this logic, because it's challenging to work
with.